+++ public/admin/_app/immutable/chunks/BOtHCHOQ.js
@@ -1,0 +1,1 @@
+import{l as e}from"./Bxz4DL7j.js";function t(t,n){throw new e(t,n.toString())}export{t};