+++ public/admin/_app/immutable/nodes/14.CJ0Q8H6x.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/nJBB0nGZ2.js";export{e as component};