+++ public/admin/_app/immutable/nodes/17.D5pitSz7.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/BSz59kwy2.js";export{e as component};