+++ public/admin/_app/immutable/nodes/18.Cbhs5H3v.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/CAEMbB0p2.js";export{e as component};