+++ public/admin/_app/immutable/nodes/19.CFN6XgUX.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/7o59LnJg.js";export{e as component};