+++ public/admin/_app/immutable/nodes/22.C1Jp28n6.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/D0Kb_yNd.js";export{e as component};