+++ public/admin/_app/immutable/nodes/23.Cck_g5v8.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/RFbpl5xe.js";export{e as component};