+++ public/admin/_app/immutable/nodes/24.DNLWIHaT.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/GvkwTSkE.js";export{e as component};